fix: fix bug with no displaying tables in PGadmin
