+++ src/db.js
@@ -1,0 +1,8 @@
+const Sequelize = require('sequelize')
+
+const db = new Sequelize('mydb', 'postgres', '54321', {
+  host: 'localhost',
+  dialect: 'postgres'
+})
+
+module.exports = db
