+++ src/fakeData/fakeData.json
@@ -2,19 +2,19 @@
   {
     "login": "blaumeux@gmail.com",
     "password": "qwert56",
-    "age": "12",
+    "age": 12,
     "isDeleted": true
   },
   {
     "login": "billy@gmail.com",
     "password": "qweewqawd",
-    "age": "13",
+    "age": 13,
     "isDeleted": true
   },
   {
     "login": "servey@gmail.com",
     "password": "qwert56dd",
-    "age": "18",
+    "age": 18,
     "isDeleted": true
   },
   {
@@ -26,7 +26,7 @@
   {
     "login": "lucan@gmail.com",
     "password": "asefff12bd",
-    "age": "23",
+    "age": 23,
     "isDeleted": true
   }
 ]
