--- src/sequelize/models/db.cjs
@@ -1,8 +1,0 @@
-var Sequelize = require('sequelize')
-
-var db = new Sequelize('mydb', 'postgres', '54321', {
-  host: 'localhost',
-  dialect: 'postgres'
-})
-
-module.exports = db
